Basic main
Finally got a working game loop. The sorting algorithms aren't fully implemented yet, but you can play against an AI that makes completely random guesses. Input validation is currently not implemented for the game loop (An invalid input will break the program), but I'm working on that right now. Just be careful what you type and the program should function as intended. The loop itself could probably use some refactoring because it's nearly 300 lines long.

The program uses one cast for the entire runtime because for some reason it didn't like it when I re-generated it.

Just let me know how the algorithms are supposed to be used, and I can work that in.